requirements.txt
python-telegram-bot==13.7
python-dotenv==0.19.1
redis==3.5.3
vk-api==11.9.6